Getting rid of `definedBefore` in `CallResolver` (#458)
Getting rid of definedBefore in `CallResolver`

The CPG aims to be a fuzzy parser and not a compiler. We therefore have more leeway when parsing declarations than a compiler. We therefore, explictily should NOT check, whether a declaration is really before its usage. The simple reason for that is that we want to be able to analyze code fragments, i.e. if the definition was in a header file that we did not see.

Closes #457

Co-authored-by: Konrad Weiss <konrad.weiss@aisec.fraunhofer.de>