Smart guessing of a cast expression vs. call expression, if the type is known
C/C++ has some ambuigities, for example in the expression `(size_t)(42)`. If CDT has not seen the type, it will guess, that this is a call expression. We try to be a little bit smarter than CDT and check our type manager, if the type is already known, for example because it was used as a type in a LHS of an assignment. This would also allow us to pre-fill the `TypeManager` with some common types.

Fixes #474