Identify and preprocess selected incremental changes
When a new version of a file has been parsed, its AST is compared with the previous version. Some of the changes, like changing a field initializer, can be applied to the CPG without having to run all passes again. In such a case, preprocessing handlers take care of integrating the change. More complex changes still trigger the execution of all passes such that a correct state is guaranteed.