TypeScript language frontend (#462)
This PR adds *experimental* support for a TypeScript parser. It is definitely not feature complete, but can be used to parse simple typescript snippets through the official `typescript` parser written in TypeScript / nodejs. It includes a simple nodejs script that invokes this parser in `src/main/nodejs`. It basically dumps the AST in a JSON structure on stdout and this input is parsed by `TypeScriptLanguageFrontend`.

This frontend should be able to parse the following languages / dialects, but only TypeScript has been tested extensively:

- JavaScript (`.js`)
- JavaScript React (`.jsx`)
- TypeScript (`.ts`)
- TypeScript React (`.tsx`)

Work that still needs to be done:
- Support basic expressions, such as calls and field access
- Support for simple classes and interfaces (no inheritance)
- Very basic React support, it will generate generic `Expression`s with the JSX content
- Adjust README with instructions
- Make it build on GitHub CI
- Bundle nodejs parser when building on JitPack
- Parse function-based React components
- Support annotation parser (aka Decorators)