Fixing type propagation behaviour between `NewExpression` and `ConstructExpression` (#504)
Type propagation between NewExpression and ConstructExpression were broken. Partly introduced by #444, partly before. This PR tries to fix this with more constant behaviour. Basically, the language frontends are now in charge of properly creating the types of NewExpressions and ConstructExpression. Before that, it somehow relied on a weird mix of type propagation and a fixInitializer function. Now I clearly documented in the wiki, that the language frontend needs to take care of this. This is needed because the behaviour is different between languages with pointer and non-pointer types.

In an ideal cpg, we would have type propagation rules that could be configured per language frontend.

There is still a (weird) type propagation towards ConstructExpression, mainly from variable declarations. I will leave this for now, even though there should never be any type propagation towards a construct expression, but only from it.