Fixes to `VariableUsageResolver`
This fixes a bug in the `VariableUsageResolver`, in which the resolver was trying to resolve a variable, that was already resolved. This can lead to weird behaviours. We always had the stance, that IF the language frontend sets the `refersTo` property, the pass should respect this and not re-set it. There are some particular cases in which the language frontend just knows better.