Parse comments and locations feature to the typescript frontend (#616)
Adds Comment parsing to the Typescript language frontend.

The position of comments are not provided by the node parser itself. It was therefore necessary to read them out manually, here I had to resort to a search by regex. Parsing comments with a regex is an imperfect solution as a context sensitive parser would be necessary. However it should be sufficient to add this feature without switching out the current parser. 
This PR:

- Computes and adds regions of the form line:column to line:column to every node, conforming to other frontends.
- Corrects a positioning error the parser makes when computing the starting position
- Parses comments from the file through a regular expression
- matches comments 

Co-authored-by: Banse, Christian <christian.banse@aisec.fraunhofer.de>