Trying to mitigate a CXX ambiguity in CDT that leads to a crash
This PR tries to mitigate #648 by making the EOG pass more robust as well as trying to recognize a certain CXX ambiguity (a call expression mistaken as a function declaration) by returning a problem declaration instead. This should at least not destroy the remaining AST scopes as it did before.