Fluent Node DSL
This PR adds a fluent node DSL to quickly build up a tree of CPG nodes. This can be used for unit tests or any occasion you quickly want to
spin up a tree of CPG nodes that are independent of the underlying programming language. This is useful if you want to test the behaviour of
some passes purely on the CPG tree, rather than a concrete program.