Introducing `FunctionType` as type for `FunctionDeclaration`
Previously, the `type` of a function was its return type. This was confusing on many levels. This PR therefore introduces a new `FunctionType`, which represents a function (and will probably some day replace `FunctionPointerType` for non-C++ languages). This will also allow us to better handle references to functions and methods in other languages that do not have the concept of a "function pointer".

Additionally, the `FunctionDeclaration` class has now a property `returnTypes`, which holds all return types. This is intentionally designed as a list, as we need to support a growing list of languages that have multiple return values (Go, Python). This is a small step towards that.