Introduction of `callee` in `CallExpression`
This PR introduces an additional field `callee` in a call expression. The idea behind this field is to have an `Expression` that is being "called" by the call expression. For most cases, this will be a `DeclaredReferenceExpression` pointing to the function declaration. However, in some languages, other things might be "callable".

Note, that this is currently not yet used in the `CallResolver`, but only introduces the field.