Trying to make the regression test slightly faster (#955)
We currently experience an issue on GitHub where sometimes (maybe because some hidden quota is hit), the GitHub runners slow down and we are not able to finish the regression test in time (20s), even though it still takes only about 2s on a good machine (aka my MacBook).

This disables parallel processing and might make it just fast enough for those special moments. Let's see.