Introduce `AssignExpression` to support multiple assignments (#1105)
This adds a new statement class, called AssignExpression as well as TupleType.

The AssignExpression holds a list of lhs and rhs for the respective values on the left and right side. Furthermore, it contains a list of declarations, since some languages can declare variables in an assignment-like statement (either implicitly or explicitly). The recommended way to fill these declarations are in an extra pass that runs after all language frontends have been executed.