Inferring missing functions in namespace declarations (#1114)
* Inferring function declarations in namespace declarations

Previously, we put all inferred function declarations in the global scope (the current TU). However, there are cases where the function is actually residing in a namespace declaration and we need to put it there.

This will not (yet) infer the namespace declaration if it is missing. This might even something that is specific to the frontend, e.g., generate namespace declarations out of import statements.

* Less functions because of less duplicates
* Added wrapper around jumpTo
* Made jumpTo private