Made the `FunctionPointerCallResolver` compatible with the `ControlFlowSensitiveDFGPass`
The function pointer call resolver pass had a hidden blocking dependency on the `ControlFlowSensitiveDFGPass`, in a way that it needed to be executed BEFORE the control flow sensitive DFG was built. This was/is because of inconsistencies between both the "regular" DFG pass and the control-flow sensitive one (which we need to tackle in a seperate PR).

This PR makes resolving function pointers a little bit cleaner, since basically it tries to resolve every `callee` of a `CallExpression` that is of `FunctionPointerType`, regardless of any AST structure or other weird resolving. This also moves the `VariableUsageResovler` a little bit closer to resolving not just variables, but also functions (since in the long term we want to merge it with the `CallResolver`)