Overhaul of the pass system
This PR contains a slight overhaul of the pass system. Specifically, instead of applying a pass to the whole translation result, it is applied to individual components within that result. That will allow us in the future, to have language-specific execution of passed in individual software components (e.g., server code and client code).