Add --custom-pass-list and --list-passes (#1147)
* Add --custom-pass-list and --list-passes

Adds a CLI option to cpg-neo4j that allows a user to choose specific
passes to run. Example: Use `--custom-pass-list=DFGPass` to run only
DFGPass. This is especially useful for debugging.

This feature requires TranslationConfiguration to keep a list of all
choosable passes. Currently, the list is stored in a global variable
of `KClass` objects (that reflection is used on in order to get their
names and create instances of them). For now, this variable only
contains passes that are part of `defaultPasses()`. This could be
easily changed by adding other passes to the variable.

* Miscellaneous small fixes

* Reformat

* Fix wrong member type

* Reformat (again, sorry! syntax error must have tripped up the formatter)

* Fix --list-passes as sole argument

* Moved custom pass list to cpg-neo4j; added FQDN

---------

Co-authored-by: Tobias Specht <tobias.specht@aisec.fraunhofer.de>