Only register extra passes if `defaultPasses()` was called
With the exctraction of the C/C++ frontend and the move from the function call resolver to an an extra pass, we encountered a problem, in which we always tried to execute the extra pass even though default passes were disabled.

This PR now checks, whether default passes need to be run and only register extra passes from language frontend if they are enabled.

Also fixes a minor bug in the CI pipeline