Added `TupleDeclaration` (#1285)
* Added `TupleDeclaration`

This PR adds a new type of `Declaration`, which represenst a tuple of variable declarations. This is primarily needed for languages, such as Go, which support the assignment of multiple values in an initializer. In this case we cannot assign the initializer of the individual variables, but assigned it to the `TupleDeclaration` instead.

Currently, the type of the `TupleDeclaration` and its elements MUST be set to the `AutoType`.

* Added DFG edges for TupleDeclaration

* DFGTest fails successfully

* Fix cfsensitivedfgpass

* Fix

* Extend DFG spec

* Create EOG for tupleDecl

* eog spec for tupleDecl

* Addressed some review comments

* update specs

---------

Co-authored-by: Alexander Kuechler <alexander.kuechler@aisec.fraunhofer.de>