Only store single unique `Type` nodes for each type
Instead of gathering individual type node objects and then squashing them back together, we just make sure that `registerType` only returns the single unique type object. However, the input into the function is still a `Type` object, which needs to be created temporarily (and then GC'd shortly after).

To reduce the number of `Type` nodes needed as input, the `objectType` function does a pre-check now, if a `ObjectType` with the given name already exists, so new `Type` object is even created. This only is possible for non-generics types though.