Added `registerReplacement` to `ScopedWalker` (#1403)
In some frontends (e.g., the Go frontend) we are replacing certain nodes in an extra pass, while iterating with the `ScopedWalker`. One such usecase is to replace calls with casts. Previously, when doing so, chained replacement of calls were a problem, since the original replaced value was stil registered in the walker and was used to determine further AST children to "walk" on, instead of the new node.

This adds a function to register such replacements with the walker.

Co-authored-by: Konrad Weiss <konrad.weiss@aisec.fraunhofer.de>