Inference of global variables
Currently, the `Inference` class only infers fields. However, some languages (such as C/C++) have global variables, which we can also infer.

Note, that we intentionally *NOT* infer local variables, since we assume that we have incomplete code, that is probably complete within a function, but not complete in the sense that translation unit files or headers are missing.