Fixing record declaration inference for pointer types.
This PR changes the type that is forwarded to the inference type. Previously any type (including a pointer type) was forwarded to `inferRecordDeclaration`. This was not working for pointer types. Now instead, the `root` type is used, which points to the element type in case of (chained) pointer types.

Also changed an error message in the scope manager to a warning and improved logging in the inference system slightly.

Fixes #1475