Extracted call/cast replacement into separate pass (#1499)
* Extracted call/cast replacement into separate pass

Multiple languages, such as Go and C++ support "functional style casts", in the form of int(5). During the frontend parsing, they are indistinguishable from function calls. Therefore, we need to do a cleanup after all types are known but before other passes are invoked that replace those calls with casts. The proposed solution is to include a new language trait HasFuntionalCasts and to move the logic from the GoExtraPass into a separate, language-neutral one.

Fixes #1487

* Added annotation to pass that requires a language trait