Fixed crash in CXXExtraPass::convertOperators
This is more a band-aid, than an actually "healing". The issue is that we often use the `language` to get a translation context in passes. But (sometimes) the language does not have it, since it is created using `registerLanguage` even before a translation context exists.

In an attempted fix, I am now using the translation context of a different node, but we should really think about how to deal with that in the future.

Fixes #1617