Populating `codeOf` of `ASTModule` (#1634)
* Populating `codeOf` of `ASTModule`

This error message was annoying me, so here is a fix. I am not sure if it is easy to find the "location" of the module. The region should be able to be computed out of the lengt of `fileContents` and its actual content. I actually thought we had a function like that somewhere @konradweiss?

* make sonarcube happy

---------

Co-authored-by: Maximilian Kaul <maximilian.kaul@aisec.fraunhofer.de>