Demote frontend dependencies to `implementation` (#1926)
Currently, the depenendency injected frontends are `api` dependencies. That means that published artefacts that dynamically depend on the configured frontends (such as Neo4J) publish all frontends as a dependency. That was not the intention. Instead they should contain no external dependencies to frontends and rather use the frontends that are available when the user loads them in.