Small improvements in `TypeResolver`
This PR sets the debug messages about type resolving from debug to trace, since they are ALOT in a real-world code base. Furthermore, `ScopeManager::lookupUniqueTypeSymbolByName` will return the first result instead of null now in case of a non-unique result. The issue with returning null is that this makes the type system think that the type does not exist and it tries to infer it, which creates more ambiguous types in the end.