Using `UnreachableEOGPass` in python
This PR adds a special version of the `UnreachableEOGPass` to the python frontend. It is executed before the symbol resolver and uses the python value evaluator. However, it is not completly usable yet, because the unreachable EOG pass is very slooooow for large code-bases. So this is blocked until #1811 is resolved.