Support different top levels for each component (#1960)
* Support different top levels for each component

This PR introduces the possibility to specify *one* top level *per* component instead of just a global one. This introduces a new `topLevels` builder in the `TranslationConfiguration`. The single `topLevel` flag is still there and sets the top level of the default component ("application"). The top level can be fetched via a virtual property `topLevel` in the `Component` class.

I also added a python test that parses different components to demonstrate the API.

Fixes #1959

* Cleanup of TranslationResult::addTranslationUnit

* Log error if application component is not there