Make the `SymbolResolver` more extensible
This PR makes all the necessary handle functions `open` so that someone external to this library can extend and replace this pass, if needed. Furthermore, we make all functions we do NOT wish to be overriden `private` (mostly because we consider them part of a legacy API).

Furthermore, this PR adds more documentation to this pass.