Using translation unit of source as start instead of global scope `astNode`
Currently, we use the global scope's `astNode` as a start node for inference of "global" things. However, this is very buggy, since the `astNode` of the global scope is assigned to the last TU we parse. If this TU was parsed in another language, we can run into situations where we infer things in the wrong language. This fixes that by using the source's TU as the inference start. This will then correctly use the sources langauge.