Attaching configuration operation to "sub"-configuration item instead of the whole config
This makes the API a little bit cleaner since we do not need to pass the conf object, which is already known and it attaches the operations to their respective counterparts, e.g. reading an option is an operation on an option now and not of the overall configuraton. It also introduces `allOps` as a way to get all the operations of the whole config again.