Fixed issues with parsing additional sources
This PR fixes some issues with parsing of additional sources. The main problem was that if include parts are "sister" paths, e.g., ../external/lib1 and ../external/lib2, then the previous logic return an incorrect relativ path of ../lib2 instead of "lib2". This PR changes that and optimizes the implementation in a way that the additional sources are already prepared in a relative way in the translation manager (the original include path is kept to be sure). This way, a language frontend can just focus on looking at relative paths and see whether they match for example a package structure.