Changing function signature of `parse` to accept the file content instead of a file (#1706)
* Changing function signature of `parse` to accept the file content instead of a file

This PR changes the way `parse` works (in a backwards compatible way). Instead of parsing a `File`, we parse the file contents (and a path). The reasoning behind this is that almost all language frontends currently need to read the file contents and we can harmonize this. This will also allow us to provide more common statistics about the parsing context in the future.

* Fixed spotless

* fixing uri handling due to it being nullable now

* Fixing parameter order

* Addign a Converter for the newly introduced TranslationStats

---------

Co-authored-by: Konrad Weiss <konrad.weiss@aisec.fraunhofer.de>